[canary][plaster] Modularise some of Brockit for reuse
This PR breaks some of the components of brockit into their own files
for reuse in a plaster prototype.

Resolves https://github.com/brave/brave-browser/issues/45113
